power actor state
