+++ core/api/full_node/make.cpp
@@ -27,8 +27,8 @@
 #include "vm/actor/builtin/states/init/init_actor_state.hpp"
 #include "vm/actor/builtin/states/market/market_actor_state.hpp"
 #include "vm/actor/builtin/states/miner/miner_actor_state.hpp"
-#include "vm/actor/builtin/states/power_actor_state.hpp"
 #include "vm/actor/builtin/states/reward/reward_actor_state.hpp"
+#include "vm/actor/builtin/states/storage_power/storage_power_actor_state.hpp"
 #include "vm/actor/builtin/states/verified_registry_actor_state.hpp"
 #include "vm/actor/builtin/types/market/deal.hpp"
 #include "vm/actor/builtin/types/market/policy.hpp"
