+++ core/vm/actor/builtin/states/all_states.hpp
@@ -5,10 +5,6 @@
 
 #pragma once
 
-#include "vm/actor/builtin/v0/storage_power/storage_power_actor_state.hpp"
-#include "vm/actor/builtin/v2/storage_power/storage_power_actor_state.hpp"
-#include "vm/actor/builtin/v3/storage_power/storage_power_actor_state.hpp"
-
 #include "vm/actor/builtin/v0/system/system_actor_state.hpp"
 #include "vm/actor/builtin/v2/system/system_actor_state.hpp"
 #include "vm/actor/builtin/v3/system/system_actor_state.hpp"
