+++ core/vm/actor/builtin/v0/storage_power/CMakeLists.txt
@@ -4,11 +4,11 @@
 #
 
 add_library(storage_power_actor_v0
-    storage_power_actor_state.cpp
     storage_power_actor_export.cpp
     storage_power_actor_utils.cpp
     )
 target_link_libraries(storage_power_actor_v0
+    storage_power_actor_state
     actor
     alpha_beta_filter
     block
