+++ core/vm/actor/builtin/v2/storage_power/CMakeLists.txt
@@ -5,9 +5,9 @@
 
 add_library(storage_power_actor_v2
     storage_power_actor_export.cpp
-    storage_power_actor_state.cpp
     storage_power_actor_utils.cpp
     )
 target_link_libraries(storage_power_actor_v2
+    storage_power_actor_state
     storage_power_actor_v0
     )
