+++ core/vm/actor/builtin/v2/storage_power/storage_power_actor_export.cpp
@@ -6,7 +6,7 @@
 #include "vm/actor/builtin/v2/storage_power/storage_power_actor_export.hpp"
 
 #include "common/logger.hpp"
-#include "vm/actor/builtin/states/power_actor_state.hpp"
+#include "vm/actor/builtin/states/storage_power/storage_power_actor_state.hpp"
 #include "vm/actor/builtin/v2/init/init_actor.hpp"
 #include "vm/actor/builtin/v2/miner/miner_actor.hpp"
 #include "vm/actor/builtin/v2/reward/reward_actor.hpp"
