+++ core/vm/actor/builtin/v3/storage_power/CMakeLists.txt
@@ -7,5 +7,6 @@ add_library(storage_power_actor_v3
     storage_power_actor_export.cpp
     )
 target_link_libraries(storage_power_actor_v3
+    storage_power_actor_state
     storage_power_actor_v2
     )
