+++ core/vm/actor/builtin/v3/storage_power/storage_power_actor_export.hpp
@@ -7,7 +7,6 @@
 
 #include "vm/actor/actor_method.hpp"
 #include "vm/actor/builtin/v2/storage_power/storage_power_actor_export.hpp"
-#include "vm/actor/builtin/v3/storage_power/storage_power_actor_state.hpp"
 
 namespace fc::vm::actor::builtin::v3::storage_power {
