+++ core/vm/runtime/circulating.cpp
@@ -9,8 +9,8 @@
 #include "const.hpp"
 #include "primitives/block/block.hpp"
 #include "vm/actor/builtin/states/market/market_actor_state.hpp"
-#include "vm/actor/builtin/states/power_actor_state.hpp"
 #include "vm/actor/builtin/states/reward/reward_actor_state.hpp"
+#include "vm/actor/builtin/states/storage_power/storage_power_actor_state.hpp"
 #include "vm/actor/builtin/types/miner/policy.hpp"
 #include "vm/state/impl/state_tree_impl.hpp"
